RetrieveCertificate now resets while retrieving
When enrolling a new certificate, for example by submitting a
user-provided CSR, people using cert-manager and other tools that
automatically enroll certificates would get "stuck" with the error of
the like:

    500 Certificate \VED\Policy\Test\foo.com has encountered an error while processing,
    Status: This certificate cannot be processed while it is in an error state. Fix any
    errors, and then click Retry., Stage: 700.

(the "stage" number doesn't matter)

Running "vcert enroll" or forcing the re-issuance in cert-manager would
not have the expected effect: the same error would show up over and
over.

The initial idea was to do a systematic "reset" before requesting, but
it was deemed too costly, as it would mean one more HTTP call on the
happy path of "vcert enroll".

The proposed solution/workaround is to do that in the
RetrieveCertificate func. It isn't ideal, but it should not affect
people since /retrieve and /reset share the same OAuth scope.